pyproject.toml
[tool.poetry]
name = "photobox-backend-grpc"
version = "0.1.0"
description = ""
authors = ["Dario Breitenstein <dario@imakethings.ch>"]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.51.1"
protobuf = "^4.21.12"
gphoto2 = "^2.3.4"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
